feat: front-matter and interpolated vars support
build script now parses front-matter of the markdown documents,
it does this in the marked preprocessing step and stores the map
on the global marked instance. This is convenient because we have access
to it in the postprocessing step when these variables are used, and they
are overwritten each time a markdown file is parsed. The caveat is
that this would cause race consitions if the script were to be modified
to run the processing in paralell, but in that case it can just be rewritten.

vars in the html file ({{var}}) can be interpolated and injected
based on matching data in the front-matter.